+++ tests/pytest.ini
@@ -1,4 +1,3 @@
 [pytest]
 markers =
-    CPU:Run the CPU-only tests
-    local:Run the local deployment tests
+    deepspeed:Run test for deepspeed CI
