+++ tests/test_non_persistent_deployment.py
@@ -2,66 +2,12 @@
 # SPDX-License-Identifier: Apache-2.0
 
 # DeepSpeed Team
+
 import pytest
-import os
-from types import SimpleNamespace
-from .utils import *  # noqa: F401
 import mii
 
 
-@pytest.fixture(scope="function")
-def mii_configs(
-    tmpdir: str,
-    dtype: str,
-    tensor_parallel: int,
-    load_with_sys_mem: bool,
-):
-
-    return {
-        'dtype': dtype,
-        'tensor_parallel': tensor_parallel,
-        'load_with_sys_mem': load_with_sys_mem,
-    }
-
-
-@pytest.fixture(scope="function")
-def deployment_config(task_name: str,
-                      model_name: str,
-                      mii_configs: dict,
-                      enable_deepspeed: bool,
-                      enable_zero: bool,
-                      ds_config: dict):
-    config = SimpleNamespace(task=task_name,
-                             model=model_name,
-                             deployment_type=mii.DeploymentType.NON_PERSISTENT,
-                             deployment_name=model_name + "_deployment",
-                             model_path=os.getenv("TRANSFORMERS_CACHE",
-                                                  None),
-                             mii_config=mii_configs,
-                             enable_deepspeed=enable_deepspeed,
-                             enable_zero=enable_zero,
-                             ds_config=ds_config)
-    validate_config(config)
-    return config
-
-
-@pytest.fixture(scope="function")
-def non_persistent_deployment(deployment_config, expected_failure):
-    if expected_failure is not None:
-        with pytest.raises(expected_failure) as excinfo:
-            mii.deploy(**deployment_config.__dict__)
-        yield excinfo
-    else:
-        mii.deploy(**deployment_config.__dict__)
-        yield deployment_config
-        mii.terminate(deployment_config.deployment_name)
-
-
-''' Unit tests '''
-
-
-@pytest.mark.local
-@pytest.mark.parametrize("dtype", ['fp16', 'fp32'])
+@pytest.mark.parametrize("deployment_type", [mii.DeploymentType.NON_PERSISTENT])
 @pytest.mark.parametrize(
     "task_name, model_name, query",
     [
@@ -119,7 +65,7 @@ def non_persistent_deployment(deployment_config, expected_failure):
         ),
     ],
 )
-def test_single_GPU(non_persistent_deployment, query):
-    generator = mii.mii_query_handle(non_persistent_deployment.deployment_name)
+def test_single_GPU(deployment, query):
+    generator = mii.mii_query_handle(deployment.deployment_name)
     result = generator.query(query)
     assert result
